# api-doc.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: VinDr Lab API Document
  contact:
    name: VinDr Lab Development Team
  license:
    name: MIT
paths:
...
  /objects:
    get:
      operationId: fetchObjects
      parameters:
        - $ref: "#/components/parameters/authParam"
      responses:
        "200":
          description: get studies response
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Object"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      description: create an Object
      operationId: createObject
      parameters:
        - $ref: "#/components/parameters/authParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObjectBig"
      responses:
        "200":
          description: get response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /objects/{object_id}:
    put:
      description: update Object
      operationId: updateObject
      parameters:
        - $ref: "#/components/parameters/authParam"
        - name: object_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Object"
      responses:
        "200":
          description: update Object
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - properties:
                      data:
                        type: object
                        properties:
                          _id:
                            type: string
                            format: uuid
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    authParam:
      name: Authorization
      in: header
      schema:
        type: string
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          enum:
            - value: 0
              name: Success
            - value: -1
              name: Failed
            - value: -2
              name: Invalid Data
            - value: -3
              name: Server Error
        message:
          type: string
        count:
          type: integer
          format: int32
          description: size of returned data
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        creator_id:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        assignee_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [NEW, DOING, COMPLETED]
        created:
          type: integer
          format: int64
        modified:
          type: integer
          format: int64
        type:
          type: string
          enum: [ANNOTATE, REVIEW]
        comment:
          type: string
        archive:
          description: default value is false
          type: boolean
...
    Object:
      type: object
      description: |
        if type=STUDY: study_instance_uid != nil, series_instance_uid = nil, sop_instance_uid = nil
        if type=SERIES: study_instance_uid != nil, series_instance_uid != nil, sop_instance_uid = nil
        if type=IMAGE: study_instance_uid != nil, series_instance_uid != nil, sop_instance_uid != nil
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [STUDY, SERIES, IMAGE]
        project_id:
          type: string
          format: uuid
        study_id:
          type: string
          format: uuid
        meta:
          type: object
          properties:
            study_instance_uid:
              type: string
              format: uuid
            series_instance_uid:
              type: string
              format: uuid
            sop_instance_uid:
              type: string
              format: uuid
    ObjectBig:
      type: object
      properties:
        project_id:
          type: string
          format: uuid
        study_id:
          type: string
          format: uuid
        list_study_instance_uid:
          type: array
          items:
            type: string
            format: uuid
        list_series_instance_uid:
          type: array
          items:
            type: string
            format: uuid
        list_sop_instance_uid:
          type: array
          items:
            type: string
            format: uuid
...